requirements.txt
llama-index-llms-ollama==0.9.1
llama-index-core==0.14.12
dotenv==0.9.9